Updated the timer constructor, Added the Character class, now seeing if I can get the world to draw the character

+++ src/util/Timer.cpp
@@ -1,6 +1,6 @@
 #include "Timer.hpp"
 
-Timer::Timer(unsigned int ms) {
+Timer::Timer(int64_t ms) {
     setTargetMs(ms);
     reset();
 }
